+++ pdf_pilot/integrations/__init__.py
@@ -1,0 +1,17 @@
+"""LangChain / LlamaIndex integrations for pdf_pilot"""
+
+from pdf_pilot.integrations.langchain import (
+    to_langchain_document,
+    to_langchain_documents,
+)
+from pdf_pilot.integrations.llamaindex import (
+    to_llamaindex_document,
+    to_llamaindex_documents,
+)
+
+__all__ = [
+    "to_langchain_document",
+    "to_langchain_documents",
+    "to_llamaindex_document",
+    "to_llamaindex_documents",
+]